requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
flask-ngrok==0.0.25
google-api-core==2.10.2
google-auth==2.14.0
google-cloud-texttospeech==2.12.3
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
playsound==1.3.0
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pyngrok==5.1.0
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
twilio==7.15.1
urllib3==1.26.12
Werkzeug==2.2.2
